refactor(api): extract shared webhook inbound helpers
Move body/secret/load/deliver shared by bearer and GitHub routes into
webhook-inbound.ts, keep GitHub route ownership in github-webhook.ts, and
cover duplicate X-GitHub-Delivery clientNonce reuse.
